requirements.txt
sqlalchemy
psycopg2-binary
requests
schedule
python-dotenv
sentry-sdk
